requirements.txt
scipy
numpy
Pillow
PyQt5
os-sys>=2.0.0
opencv-python
python-math
pyglet
matplotlib
pytest-timeit